requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==2.1.1
Django==3.1.3
django-allauth==0.42.0
django-currentuser==0.5.2
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
